fix standalone build: drop Decoder.CanDecode, clean RVA assembly
- iced-x86 Decoder has no CanDecode property; decode until decoder.IP
  reaches the function's .pdata end address instead.
- Assemble the 32-bit RVA from unsigned bytes to avoid the CS0675
  sign-extension warning.

+++ standalone/PeImage.cs
@@ -210,7 +210,7 @@ namespace OffsetRecovery.Standalone
             {
                 return -1;
             }
-            return b0 | (b1 << 8) | (b2 << 16) | ((long)b3 << 24);
+            return (uint)b0 | ((uint)b1 << 8) | ((uint)b2 << 16) | ((uint)b3 << 24);
         }
     }
 }
